Refactor DigitiserReader
