+++ trace-reader/src/hdf5trace/mod.rs
@@ -3,24 +3,21 @@ mod channel;
 mod digitiser;
 mod read_engine;
 
-use crate::{Hdf5, hdf5trace::read_engine::ReadCommand};
+use crate::{
+    Hdf5,
+    hdf5trace::read_engine::{DigitiserReader, ReadCommand},
+};
 use chrono::ParseError;
 use digital_muon_common::{
     DigitizerId, FrameNumber,
     spanned::{SpanWrapper, Spanned},
 };
-use digital_muon_streaming_types::flatbuffers::FlatBufferBuilder;
 use hdf5::{File, OpenMode};
 use rayon::iter::{IntoParallelRefMutIterator, ParallelIterator};
-use rdkafka::{
-    ClientConfig,
-    error::KafkaError,
-    producer::{BaseRecord, DefaultProducerContext, Producer, ThreadedProducer},
-    util::Timeout,
-};
-use std::{fmt::Debug, num::ParseIntError, ops::Range, path::PathBuf, str::FromStr};
+use rdkafka::{ClientConfig, error::KafkaError};
+use std::{fmt::Debug, num::ParseIntError, path::PathBuf, str::FromStr};
 use thiserror::Error;
-use tracing::{debug, error, info, info_span};
+use tracing::{debug, info_span};
 
 pub(crate) use digitiser::{HDF5Config, Hdf5Digitiser};
 
@@ -146,7 +143,7 @@ pub(crate) async fn read_hdf5_file(
 
     let digitiser_present = digitisers
         .iter()
-        .map(|d| d.digitiser.get_id())
+        .map(|d| d.digitiser().get_id())
         .collect::<Vec<_>>();
 
     let mut digitisers = digitisers
@@ -155,20 +152,14 @@ pub(crate) async fn read_hdf5_file(
         .collect::<Vec<_>>();
 
     if args.summary_only {
-        for digitiser in digitisers.iter_mut() {
-            digitiser.digitiser.output_summary();
-        }
+        /*for digitiser in digitisers.iter_mut() {
+            //digitiser.digitiser().output_summary();
+        }*/
     } else {
         for command_index in 0..read_sequence.len() {
             let num_indices = digitisers
                 .iter()
-                .map(|digitiser| {
-                    digitiser
-                        .read_sequence
-                        .get(command_index)
-                        .expect("This should never fail.")
-                        .len()
-                })
+                .map(|digitiser| digitiser.get_command(command_index).len())
                 .min()
                 .ok_or_else(|| Error::NoDigitisersSelected(digitiser_present.clone()))?;
             for index in 0..=num_indices {
@@ -186,139 +177,6 @@ pub(crate) async fn read_hdf5_file(
     }
 
     Ok(())
-}
-
-/// Encapsulates the tools needed to read the digitiser messages from a hdf5 file and produce them to the kafka broker.
-struct DigitiserReader {
-    /// Encapsulates the metadata and link to the hdf5 file for the digitiser messages.
-    digitiser: Hdf5Digitiser,
-    /// The sequence of read_instructions to run through
-    read_sequence: Vec<Range<usize>>,
-    /// The kafka producer this digitiser uses.
-    producer: ThreadedProducer<DefaultProducerContext>,
-}
-
-impl DigitiserReader {
-    /// Creates a new instance.
-    ///
-    /// # Parameters
-    /// - client_config: the kafka config settings to use for the produer.
-    /// - args: the cli args specific to `hdf5` mode.
-    /// - digitiser:
-    fn new(
-        client_config: &ClientConfig,
-        read_sequence: &[ReadCommand],
-        digitiser: Hdf5Digitiser,
-    ) -> Result<Self, Error> {
-        let read_sequence = read_sequence
-            .iter()
-            .map(|command| match command {
-                &ReadCommand::FrameRange(from, to) => Ok(digitiser
-                    .get_index_from_frame_number(from)?
-                    ..digitiser.get_index_from_frame_number(to)?),
-                &ReadCommand::FrameCount(from, count) => Ok(digitiser
-                    .get_index_from_frame_number(from)?
-                    ..(digitiser.get_index_from_frame_number(from)?) + count),
-                &ReadCommand::IndexRange(from, to) => Ok(from..to),
-                &ReadCommand::IndexCount(from, count) => Ok(from..(from + count)),
-                ReadCommand::TimestampRange(_from, _to) => unimplemented!(),
-                ReadCommand::TimestampCount(_from, _count) => unimplemented!(),
-                ReadCommand::All => Ok(0..digitiser.get_num_frames()),
-            })
-            .collect::<Result<Vec<_>, Error>>()?;
-
-        info!(
-            "Reader for digitiser {} read command sequence : {read_sequence:?}.",
-            digitiser.get_id()
-        );
-        let producer = client_config.create()?;
-        Ok(Self {
-            digitiser,
-            read_sequence,
-            producer,
-        })
-    }
-
-    /// Read the digitiser message at the given index and produce it to the broker.
-    ///
-    /// # Parameters
-    /// - trace_topic: the Kafka topic to produce to.
-    /// - key: the text to use for the produced message's key.
-    /// - args: the cli args specific to `hdf5` mode.
-    /// - index: the index of the message to read.
-    fn read_at_index(
-        &self,
-        trace_topic: &str,
-        key: &str,
-        args: &Hdf5,
-        command_index: usize,
-        index: usize,
-    ) -> Result<(), Error> {
-        let mut fbb = FlatBufferBuilder::new();
-        self.digitiser.create_message(
-            &mut fbb,
-            self.read_sequence
-                .get(command_index)
-                .expect("This should never fail")
-                .start
-                + index,
-            args.sample_rate,
-            &args.overwrite_fields,
-        )?;
-        info_span!("Send").in_scope(|| self.send_record(&mut fbb, trace_topic, key));
-        Ok(())
-    }
-
-    /// Sends the FlatBuffer payload to the desired Kafka topic.
-    ///
-    /// # Parameters
-    /// - fbb: mutable reference to the FlatBufferBuilder to use.
-    /// - trace_topic: the Kafka topic to produce to.
-    /// - key: the text to use for the produced message's key.
-    fn send_record(&self, fbb: &mut FlatBufferBuilder, trace_topic: &str, key: &str) {
-        let base_record = BaseRecord::to(trace_topic)
-            .payload(fbb.finished_data())
-            .key(key);
-
-        let mut result = self.producer.send(base_record);
-        while let Err((_, base_record)) = result {
-            result = self.producer.send(base_record);
-        }
-    }
-
-    fn is_id_contained_in(&self, ids: &[DigitizerId]) -> bool {
-        if ids.is_empty() {
-            true
-        } else {
-            ids.contains(&self.digitiser.get_id())
-        }
-    }
-
-    /// Given an index, ensure the necessary data is in the cache.
-    /// This should each time before the `create_message` method is used.
-    ///
-    /// This method is idempotent, so does nothing if the required index is already cached.
-    ///
-    /// # Parameters
-    /// - index: the index to ensure is cached.
-    #[tracing::instrument(skip_all)]
-    pub(crate) fn ensure_elements_cached(&mut self, command_index: usize, index: usize) {
-        self.digitiser.ensure_elements_cached(
-            self.read_sequence
-                .get(command_index)
-                .expect("This should never fail.")
-                .start
-                + index,
-        );
-    }
-}
-
-impl Drop for DigitiserReader {
-    fn drop(&mut self) {
-        if let Err(e) = self.producer.flush(Timeout::Never) {
-            error!("{e}");
-        }
-    }
 }
 
 /// Read the messages at the given index, in the given slice of `DigitiserReaders` and produce them to the broker.
@@ -369,7 +227,10 @@ mod tests {
     use crate::OverwriteFields;
 
     use super::*;
-    use digital_muon_streaming_types::dat2_digitizer_analog_trace_v2_generated::root_as_digitizer_analog_trace_message;
+    use digital_muon_streaming_types::{
+        dat2_digitizer_analog_trace_v2_generated::root_as_digitizer_analog_trace_message,
+        flatbuffers::FlatBufferBuilder,
+    };
     use std::{fs::File, io::Read};
 
     #[test]
